Add files via upload
Almost there for Fracas: can't do ellipsis or "awarded and was awarded", because "was awarded" isn't a sensible phrase. Nice challenge, but not one I can see a way to taking on. A few others that need some debugging to get the right answer, but I do get something for all the rest.