Notes for Ghadah, minor patches.
notes-20-10-2017.txt is just running instructions. setup didn't include pstree.pl, which was stupid. Minor patches to chart (so that latexParse doesn't necessarily actually produce LaTeX output) and nf.pl so that Skolems look like #23 and contexts use \IN (which comes out OK when we do pretty, which is nice)